The concept of player has been added (IA and Human), with the concept of fof (friend of foe) There has been some refactoring in the entity data model. The attack action now has the ship follow the target.

+++ src/main/java/net/carmgate/morph/model/entities/orders/TakeDamageOrder.java
@@ -2,13 +2,13 @@ package net.carmgate.morph.model.entities.orders;
 
 public class TakeDamageOrder implements Order {
 
-	private final int damageAmount;
+	private final float damageAmount;
 
-	public TakeDamageOrder(int damageAmount) {
+	public TakeDamageOrder(float damageAmount) {
 		this.damageAmount = damageAmount;
 	}
 
-	public int getDamageAmount() {
+	public float getDamageAmount() {
 		return damageAmount;
 	}
